修正LocalDateTime返回数据格式

+++ src/main/java/com/yhm/universityhelper/entity/po/Task.java
@@ -14,7 +14,6 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 
@@ -80,8 +79,8 @@ public class Task implements Serializable, Comparable<Object> {
     private Integer taskState;
 
     @TableField("takeoutId")
-    private BigInteger takeoutId;
-
+    private Integer takeoutId;
+    
     @TableField("orderTime")
     private LocalDateTime orderTime;
